analytics/packages/monitoreo/requirements.txt
apache-beam[gcp]>=2.30.0
beam-nuggets==0.18.1
google-cloud-secret-manager==2.8.0
google-api-core[grpc]<2.0.0dev,>=1.22.2
google-auth<2.0dev,>=1.24.0
numpy==1.20.3
obspy==1.2.2
psycopg2-binary==2.9.2
pyparsing<3,>=2.4.2
python-dotenv==0.19.2
